requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
django-cors-headers>=4.0,<5
whitenoise>=6,<7
gunicorn>=21,<22
dj-database-url>=2,<3
